fix(batch-routing): add mobile search screen for batch routing
