+++ lib/components/app/batch-routing-panel.js
@@ -5,77 +5,12 @@ import styled from 'styled-components'
 
 import * as apiActions from '../../actions/api'
 import * as formActions from '../../actions/form'
-import BatchSettingsPanel from '../form/batch-settings-panel'
+import BatchSettings from '../form/batch-settings'
 import LocationField from '../form/connected-location-field'
-import DateTimeModal from '../form/date-time-modal'
-import ModeButtons, {MODE_OPTIONS, StyledModeButton} from '../form/mode-buttons'
-// import UserSettings from '../form/user-settings'
-import Icon from '../narrative/icon'
 import NarrativeItineraries from '../narrative/narrative-itineraries'
-import {
-  BatchSettingsPanelContainer,
-  DateTimeModalContainer,
-  Dot,
-  MainSettingsRow,
-  PlanTripButton,
-  SettingsPreview,
-  StyledDateTimePreview
-} from './styled'
-import { hasValidLocation, getActiveSearch, getShowUserSettings } from '../../util/state'
+import { getActiveSearch, getShowUserSettings } from '../../util/state'
 import ViewerContainer from '../viewers/viewer-container'
 
-/**
- * Simple utility to check whether a list of mode strings contains the provided
- * mode. This handles exact match and prefix/suffix matches (i.e., checking
- * 'BICYCLE' will return true if 'BICYCLE' or 'BICYCLE_RENT' is in the list).
- *
- * FIXME: This might need to be modified to be a bit looser in how it handles
- * the 'contains' check. E.g., we might not want to remove WALK,TRANSIT if walk
- * is turned off, but we DO want to remove it if TRANSIT is turned off.
- */
-function listHasMode (modes, mode) {
-  return modes.some(m => mode.indexOf(m) !== -1)
-}
-
-function combinationHasAnyOfModes (combination, modes) {
-  return combination.mode.split(',').some(m => listHasMode(modes, m))
-}
-
-// List of possible modes that can be selected via mode buttons.
-const POSSIBLE_MODES = MODE_OPTIONS.map(b => b.mode)
-
-const ModeButtonsFullWidthContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 5px;
-`
-
-// Define Mode Button styled components here to avoid circular imports. I.e., we
-// cannot define them in styled.js (because mode-buttons.js imports buttonCss
-// and then we would need to import ModeButtons/StyledModeButton from that file
-// in turn).
-const StyledModeButtonsFullWidth = styled(ModeButtons)`
-  &:last-child {
-    margin-right: 0px;
-  }
-`
-
-const ModeButtonsContainerCompressed = styled.div`
-  display: contents;
-`
-
-const ModeButtonsCompressed = styled(ModeButtons)`
-  ${StyledModeButton} {
-    border-radius: 0px;
-  }
-  &:first-child {
-    border-radius: 5px 0px 0px 5px;
-  }
-  &:last-child {
-    margin-right: 5px;
-    border-radius: 0px 5px 5px 0px;
-  }
-`
 // Style for setting the top of the narrative itineraries based on the width of the window.
 // If the window width is less than 1200px (Bootstrap's "large" size), the
 // mode buttons will be shown on their own row, meaning that the
@@ -97,56 +32,8 @@ const NarrativeContainer = styled.div`
  * Main panel for the batch/trip comparison form.
  */
 class BatchRoutingPanel extends Component {
-  state = {
-    expanded: null,
-    selectedModes: POSSIBLE_MODES
-  }
-
-  _onClickMode = (mode) => {
-    const {possibleCombinations, setQueryParam} = this.props
-    const {selectedModes} = this.state
-    const index = selectedModes.indexOf(mode)
-    const enableMode = index === -1
-    const newModes = [...selectedModes]
-    if (enableMode) newModes.push(mode)
-    else newModes.splice(index, 1)
-    // Update selected modes for mode buttons.
-    this.setState({selectedModes: newModes})
-    // Update the available mode combinations based on the new modes selection.
-    const disabledModes = POSSIBLE_MODES.filter(m => !newModes.includes(m))
-    // Do not include combination if any of its modes are found in disabled
-    // modes list.
-    const newCombinations = possibleCombinations
-      .filter(c => !combinationHasAnyOfModes(c, disabledModes))
-    setQueryParam({combinations: newCombinations})
-  }
-
-  _planTrip = () => {
-    const {currentQuery, routingQuery} = this.props
-    // Check for any validation issues in query.
-    const issues = []
-    if (!hasValidLocation(currentQuery, 'from')) issues.push('from')
-    if (!hasValidLocation(currentQuery, 'to')) issues.push('to')
-    if (issues.length > 0) {
-      // TODO: replace with less obtrusive validation.
-      window.alert(`Please define the following fields to plan a trip: ${issues.join(', ')}`)
-      return
-    }
-    // Close any expanded panels.
-    this.setState({expanded: null})
-    // Plan trip.
-    routingQuery()
-  }
-
-  _updateExpanded = (type) => ({expanded: this.state.expanded === type ? null : type})
-
-  _toggleDateTime = () => this.setState(this._updateExpanded('DATE_TIME'))
-
-  _toggleSettings = () => this.setState(this._updateExpanded('SETTINGS'))
-
   render () {
-    const {config, currentQuery, mobile} = this.props
-    const {expanded, selectedModes} = this.state
+    const {mobile} = this.props
     const actionText = mobile ? 'tap' : 'click'
     return (
       <ViewerContainer className='batch-routing-panel'>
@@ -160,52 +47,7 @@ class BatchRoutingPanel extends Component {
           locationType='to'
           showClearButton={!mobile}
         />
-        <ModeButtonsFullWidthContainer className='hidden-lg'>
-          <StyledModeButtonsFullWidth
-            className='flex'
-            onClick={this._onClickMode}
-            selectedModes={selectedModes}
-          />
-        </ModeButtonsFullWidthContainer>
-        <MainSettingsRow>
-          <SettingsPreview
-            expanded={expanded === 'SETTINGS'}
-            onClick={this._toggleSettings}
-          >
-            {coreUtils.query.isNotDefaultQuery(currentQuery, config) &&
-              <Dot className='dot' />
-            }
-            <Icon type='cog' className='fa-2x' />
-          </SettingsPreview>
-          <StyledDateTimePreview
-            // as='button'
-            expanded={expanded === 'DATE_TIME'}
-            hideButton
-            onClick={this._toggleDateTime} />
-          <ModeButtonsContainerCompressed>
-            <ModeButtonsCompressed
-              className='visible-lg straight-corners'
-              onClick={this._onClickMode}
-              selectedModes={selectedModes}
-            />
-          </ModeButtonsContainerCompressed>
-          <PlanTripButton
-            onClick={this._planTrip}
-            title='Plan trip'
-          >
-            <Icon type='search' className='fa-2x' />
-          </PlanTripButton>
-        </MainSettingsRow>
-        {expanded === 'DATE_TIME' &&
-          <DateTimeModalContainer>
-            <DateTimeModal />
-          </DateTimeModalContainer>
-        }
-        {expanded === 'SETTINGS' &&
-          <BatchSettingsPanelContainer>
-            <BatchSettingsPanel />
-          </BatchSettingsPanelContainer>
-        }
+        <BatchSettings />
         {/* FIXME: Add back user settings (home, work, etc.) once connected to
             the middleware persistence.
           !activeSearch && showUserSettings &&
