+++ lib/components/form/batch-preferences.js
@@ -6,10 +6,10 @@ import { setQueryParam } from '../../actions/form'
 import { ComponentContext } from '../../util/contexts'
 import { getShowUserSettings } from '../../util/state'
 
-import { StyledBatchSettingsPanel } from './styled'
+import { StyledBatchPreferences } from './batch-styled'
 import UserTripSettings from './user-trip-settings'
 
-class BatchSettingsPanel extends Component {
+class BatchPreferences extends Component {
   static contextType = ComponentContext
 
   render () {
@@ -26,7 +26,7 @@ class BatchSettingsPanel extends Component {
         <div className='modes-panel'>
           {showUserSettings && <UserTripSettings />}
 
-          <StyledBatchSettingsPanel
+          <StyledBatchPreferences
             ModeIcon={ModeIcon}
             queryParams={query}
             supportedModes={config.modes}
@@ -108,4 +108,4 @@ const mapDispatchToProps = {
   setQueryParam
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(BatchSettingsPanel)
+export default connect(mapStateToProps, mapDispatchToProps)(BatchPreferences)
