+++ lib/components/form/batch-styled.js
@@ -1,59 +1,97 @@
-import styled, { css } from 'styled-components'
-import { DateTimeSelector, SettingsSelectorPanel } from '@opentripplanner/trip-form'
 import * as TripFormClasses from '@opentripplanner/trip-form/lib/styled'
+import { SettingsSelectorPanel } from '@opentripplanner/trip-form'
+import styled, {css} from 'styled-components'
 
-const commonButtonCss = css`
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  background: none;
-  border: 1px solid rgb(187, 187, 187);
-  border-radius: 3px;
-  font-family: inherit;
-  font-size: inherit;
-  font-weight: inherit;
-  outline-offset: -2px;
-  padding: 6px 12px;
-  text-align: center;
-  touch-action: manipulation;
-  user-select: none;
+import DateTimePreview from './date-time-preview'
+import {commonInputCss, modeButtonButtonCss} from './styled'
 
-  &.active {
-    background-color: rgb(173, 216, 230);
-    border: 2px solid rgb(0, 0, 0);
-    box-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.125);
-    font-weight: 600;
-  }
+const SHADOW = 'inset 0px 0px 5px #c1c1c1'
+
+const activeCss = css`
+  background: #e5e5e5;
+  -webkit-box-shadow: ${SHADOW};
+     -moz-box-shadow: ${SHADOW};
+          box-shadow: ${SHADOW};
+   outline: none;
 `
 
-export const commonInputCss = css`
-  background: none;
-  border: 1px solid #ccc;
-  box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
-  color: #555;
-  font-family: inherit;
-  font-weight: inherit;
-  padding: 6px 12px;
-  transition: border-color ease-in-out .15s, box-shadow ease-in-out .15s;
-`
-
-export const modeButtonButtonCss = css`
-  ${TripFormClasses.ModeButton.Button} {
-    ${commonButtonCss}
-  }
-`
-
-export const Dot = styled.div`
-  position: absolute;
-  top: -3px;
-  right: -3px;
-  width: 10px;
-  height: 10px;
+export const buttonCss = css`
+  height: 45px;
+  width: 45px;
+  margin: 0px;
+  border: 0px;
   border-radius: 5px;
-  background-color: #f00;
+  &:active {
+    ${activeCss}
+  }
 `
 
-export const StyledSettingsSelectorPanel = styled(SettingsSelectorPanel)`
+export const Button = styled.button`
+  ${buttonCss}
+`
+
+export const StyledDateTimePreview = styled(DateTimePreview)`
+  ${buttonCss}
+  background-color: rgb(239, 239, 239);
+  cursor: pointer;
+  font-size: 12px;
+  margin-right: 5px;
+  padding: 7px 5px;
+  text-align: left;
+  white-space: nowrap;
+  width: 120px;
+  ${props => props.expanded ? activeCss : null}
+`
+export const SettingsPreview = styled(Button)`
+  line-height: 22px;
+  margin-right: 5px;
+  padding: 10px 0px;
+  position: relative;
+  ${props => props.expanded ? activeCss : null}
+`
+
+export const PlanTripButton = styled(Button)`
+  background-color: #F5F5A7;
+  margin-left: auto;
+  padding: 5px;
+  &:active {
+    ${activeCss}
+    background-color: #ededaf
+  }
+`
+
+const expandableBoxCss = css`
+  background-color: rgb(239, 239, 239);
+  box-shadow: rgba(0, 0, 0, 0.32) 7px 12px 10px;
+  height: 245px;
+  border-radius: 5px 5px 5px 5px;
+  left: 10px;
+  position: absolute;
+  right: 10px;
+  z-index: 99999;
+`
+
+export const DateTimeModalContainer = styled.div`
+  ${expandableBoxCss}
+  padding: 10px 20px;
+`
+
+export const BatchPreferencesContainer = styled.div`
+  ${expandableBoxCss}
+  padding: 5px 10px;
+`
+
+export const MainSettingsRow = styled.div`
+  align-items: top;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  margin-bottom: 5px;
+`
+
+// FIXME: This is identical to StyledSettingsSelectorPanel, with a
+// couple of items set to display: none (SettingsHeader and ModeSelector)
+export const StyledBatchPreferences = styled(SettingsSelectorPanel)`
   ${modeButtonButtonCss}
 
   ${TripFormClasses.SettingLabel} {
@@ -65,6 +103,7 @@ export const StyledSettingsSelectorPanel = styled(SettingsSelectorPanel)`
     text-transform: uppercase;
   }
   ${TripFormClasses.SettingsHeader} {
+    display: none;
     color: #333333;
     font-size: 18px;
     margin: 16px 0px;
@@ -101,6 +140,7 @@ export const StyledSettingsSelectorPanel = styled(SettingsSelectorPanel)`
   }
 
   ${TripFormClasses.ModeSelector} {
+    display: none;
     font-weight: 300;
     ${TripFormClasses.ModeButton.Button} {
       box-shadow: none;
@@ -178,27 +218,5 @@ export const StyledSettingsSelectorPanel = styled(SettingsSelectorPanel)`
       height: 18px;
       max-width: 32px;
     }
-  }
-`
-
-export const StyledDateTimeSelector = styled(DateTimeSelector)`
-  margin: 0 -15px 15px;
-
-  ${TripFormClasses.DateTimeSelector.DateTimeRow} {
-    margin: 20px 0px 15px;
-    input {
-      ${commonInputCss}
-      background-color: #fff;
-      border: 0;
-      border-bottom: 1px solid #000;
-      box-shadow: none;
-      outline: none;
-      text-align: center;
-    }
-  }
-  ${TripFormClasses.ModeButton.Button} {
-    ${commonButtonCss}
-    font-size: 14px;
-    height: 35px;
   }
 `
