+++ lib/components/form/mode-buttons.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 
 import Icon from '../narrative/icon'
 import { ComponentContext } from '../../util/contexts'
-import {buttonCss} from '../app/styled'
+import {buttonCss} from './batch-styled'
 
 export const MODE_OPTIONS = [
   {
