+++ lib/components/form/settings-preview.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
-import { Dot } from '../app/styled'
+import { Dot } from './styled'
 import { mergeMessages } from '../../util/messages'
 
 class SettingsPreview extends Component {
