+++ lib/components/mobile/main.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
+import BatchSearchScreen from './batch-search-screen'
 import MobileDateTimeScreen from './date-time-screen'
 import MobileOptionsScreen from './options-screen'
 import MobileLocationSearch from './location-search'
@@ -13,6 +14,7 @@ import MobileTripViewer from './trip-viewer'
 import MobileRouteViewer from './route-viewer'
 
 import { MobileScreens, MainPanelContent, setMobileScreen } from '../../actions/ui'
+import { isBatchRoutingEnabled } from '../../util/itinerary'
 import { getActiveItinerary } from '../../util/state'
 
 class MobileMain extends Component {
@@ -43,7 +45,7 @@ class MobileMain extends Component {
   }
 
   render () {
-    const { map, title, uiState } = this.props
+    const { config, map, title, uiState } = this.props
 
     // check for route viewer
     if (uiState.mainPanelContent === MainPanelContent.ROUTE_VIEWER) {
@@ -69,8 +71,13 @@ class MobileMain extends Component {
         )
 
       case MobileScreens.SEARCH_FORM:
+        // Render batch search screen if batch routing enabled. Otherwise,
+        // default to standard search screen.
+        const SearchScreen = isBatchRoutingEnabled(config)
+          ? BatchSearchScreen
+          : MobileSearchScreen
         return (
-          <MobileSearchScreen
+          <SearchScreen
             map={map}
             newScreen={this.newScreen}
           />
@@ -112,6 +119,7 @@ class MobileMain extends Component {
 
 const mapStateToProps = (state, ownProps) => {
   return {
+    config: state.otp.config,
     uiState: state.otp.ui,
     currentQuery: state.otp.currentQuery,
     currentPosition: state.otp.location.currentPosition,
